requirements.txt
matplotlib==3.6.0
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.7.3
seaborn==0.11.2
streamlit==1.14.0
